REFACTOR :: 컴포넌트는 비즈니스 로직을 분리하지 않는다
