+++ apps/user/src/app/form/지원자정보/지원자정보.hooks.ts
@@ -37,5 +37,5 @@ export const useInput = () => {
         }));
     }, [date]);
 
-    return { setUserInfo, handleUserInfoDataChange, date, setDate };
+    return { userInfo, setUserInfo, handleUserInfoDataChange, date, setDate };
 };
