+++ apps/user/src/app/form/출신학교및학력/출신학교및학력.hooks.ts
@@ -28,8 +28,6 @@ const useInput = () => {
         setEducationInfo((prev) => ({ ...prev, [name]: value }));
     };
 
-    useEffect(() => console.log(educationInfo), [educationInfo]);
-
     return { educationInfo, setEducationInfo, handleEducationInfoDataChange };
 };
