+++ apps/user/src/components/form/FindSchoolModal/FindSchoolModal.tsx
@@ -1,13 +1,13 @@
 import useSchoolListQuery from '@/services/form/출신학교및학력/queries';
+import { School } from '@/types/form';
 import { useDebounceInput } from '@maru/hooks';
 import { color, font } from '@maru/theme';
 import { Button, Column, Row, SearchInput } from '@maru/ui';
 import Check from '@maru/ui/Icons/Check';
 import Close from '@maru/ui/Icons/Close';
-import { Dispatch, SetStateAction } from 'react';
-import { css, styled } from 'styled-components';
+import { useState, Dispatch, SetStateAction } from 'react';
 import { EducationInfo } from '../../../app/form/출신학교및학력/출신학교및학력.hooks';
-import useSchoolModalHandler from './FindSchoolModal.hooks';
+import { css, styled } from 'styled-components';
 
 interface PropsType {
     closeModal: () => void;
@@ -15,64 +15,67 @@ interface PropsType {
 }
 
 const SchoolSearchModal = ({ closeModal, setEducationInfo }: PropsType) => {
-    const { value, onChange, debouncedValue } = useDebounceInput({ initialValue: '' });
+    const [selectedSchool, setSelectedSchool] = useState({ name: '', location: '', code: '' });
+    const {
+        value: schoolSearchQuery,
+        onChange: handleSchoolSearchQueryDataChange,
+        debouncedValue: debouncedSchoolSearchQuery,
+    } = useDebounceInput({ initialValue: '' });
 
-    const schoolListQuery = useSchoolListQuery(debouncedValue);
+    const schoolListQuery = useSchoolListQuery(debouncedSchoolSearchQuery);
 
-    const { selectedSchool, handleSchoolSelect, closeSchoolModal, onComplete } =
-        useSchoolModalHandler(closeModal, setEducationInfo);
+    const handleCompleteSchoolSearch = () => {
+        const { name, location, code } = selectedSchool;
+
+        setEducationInfo((prev) => ({
+            ...prev,
+            schoolName: name,
+            schoolLocation: location,
+            schoolCode: code,
+        }));
+        closeModal();
+    };
+
+    const closeSchoolModal = () => {
+        setSelectedSchool({ name: '', location: '', code: '' });
+        closeModal();
+    };
 
     return (
         <Background>
             <StyledSchoolSearchModal>
-                <Column gap="24px">
-                    <Column gap="16px">
+                <Column gap={24}>
+                    <Column gap={16}>
                         <Row justifyContent="space-between">
                             <Title>학교 검색</Title>
                             <Close cursor="pointer" onClick={closeSchoolModal} />
                         </Row>
                         <SearchInput
-                            value={value}
-                            onChange={onChange}
+                            value={schoolSearchQuery}
+                            onChange={handleSchoolSearchQueryDataChange}
                             placeholder="학교 이름을 입력해주세요."
                         />
                     </Column>
                     <SchoolList>
-                        {schoolListQuery.data?.dataList.map(
-                            ({
-                                name,
-                                location,
-                                code,
-                            }: {
-                                name: string;
-                                location: string;
-                                code: string;
-                            }) => (
-                                <SchoolItem
-                                    key={code}
-                                    selected={selectedSchool.code === code}
-                                    onClick={() =>
-                                        handleSchoolSelect({
-                                            name,
-                                            location,
-                                            code,
-                                        })
-                                    }>
-                                    <SchoolName>
-                                        {selectedSchool.code === code && <Check />}
-                                        {name}
-                                    </SchoolName>
-                                    <SchoolRegion>{location}</SchoolRegion>
-                                </SchoolItem>
-                            ),
-                        )}
+                        {schoolListQuery.data?.dataList.map(({ name, location, code }: School) => (
+                            <SchoolItem
+                                key={code}
+                                selected={selectedSchool.code === code}
+                                onClick={() => setSelectedSchool({ name, location, code })}>
+                                <SchoolName>
+                                    {selectedSchool.code === code && <Check />}
+                                    {name}
+                                </SchoolName>
+                                <SchoolRegion>{location}</SchoolRegion>
+                            </SchoolItem>
+                        ))}
                     </SchoolList>
                 </Column>
-                <Row gap="16px" justifyContent="flex-end">
+                <Row gap={16} justifyContent="flex-end">
                     <Button option="SECONDARY" size="SMALL" onClick={closeSchoolModal}>
                         취소
                     </Button>
-                    <Button size="SMALL" onClick={onComplete}>
+                    <Button size="SMALL" onClick={handleCompleteSchoolSearch}>
                         완료
                     </Button>
                 </Row>
